digit-recognition/requirements.txt
matplotlib==3.5.1
numpy==1.22.3
pylint==2.12.2
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
Pillow==9.1.0